% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifiR-package.R
\docType{package}
\name{specifiR-package}
\alias{specifiR-package}
\title{specifiR: Quantifying Host Specificity of Microbial Community Members}
\description{
This method can be used to compare the specificity of whole communities associated with groups. Groups can include hosts, genotypes, environments, treatments, etc.
}
\keyword{internal}
