% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifiR_physeq.R
\name{specifiR_physeq}
\alias{specifiR_physeq}
\title{Quantify group specificity of entire communities for each sample}
\usage{
specifiR_physeq(
  physeq,
  groups,
  seed = 666,
  n.perm = 999,
  pval.cutoff = 0.05,
  max.ratio = 0,
  ovp.plot = FALSE,
  rm.rare.taxa = TRUE
)
}
\arguments{
\item{physeq}{Phyloseq object. Must contain a valid "otu_table" slot containing raw observation counts, and a valid "sample_data" slot containing the grouping variable.}

\item{groups}{Character vector of length 1. The quoted name of the sample_data column that has grouping information.}

\item{seed}{Numeric of length 1. Random seed for reproducibility. Default = 666.}

\item{n.perm}{Positive numeric (integer) of length 1. Number of permutations for Monte Carlo permutation test. Default = 999.}

\item{pval.cutoff}{Positive numeric of length 1, between 0 and 1. The P-value cutoff for significance. Default = 0.05.}

\item{max.ratio}{The maximum ratio of significant:insignificant P-values within a group. Taxa are grouped by the number of samples in which they occur. Groups in which no taxa have a significant p-value from the ISA will be flagged as a group of rare taxa. Only the first N occurrence groups that have this value or lower will be flagged for rare taxon removal. Default = 0. You are unlikely to want to change this value.}

\item{ovp.plot}{Logical. Should a plot of occupancy vs. p-values be generated? Default = FALSE.}

\item{rm.rare.taxa}{Logical. Should rare taxa be removed before the CWM? Default = TRUE. Set to FALSE if you want to perform the Community Weighted Mean analysis on all taxa.}
}
\value{
Named List. This returns a list with 5 elements:
community_specificity_index = The main result showing community weighted mean for each sample, communities with greater values are more specific;
taxon_specificity_index = Intermediate result (`comm_name` `iv.max` `p.value` `occurrence` `occurrence_groups` `taxon_index`), taxa that are more host specific will have a greater taxon_index;
isa_results = Intermediate results. taxon-level indicator species analysis with each given group level, and the indicator results for each taxon;
process_summary = Reports basic info on process, including the number of taxa removed due to rarity;
removed_taxa = Character vector with names of taxa removed due to rarity.
}
\description{
Function specifiR assigns a community level specificity index to each community in the data set.The output values range from 0-1, with values closer to one indicating a community with taxa that are more host specific. This version of the function is a convenience wrapper for phyloseq users.
}
\details{
SpecifiR is a function used to assign specificity values to both individual taxa and entire communities in community data sets. This method is derived from two analyses, the Indicator Species Analysis (ISA) and a Community Weighted Mean (CWM) analysis.

First, the permutation test of the ISA is used to assign a specificity index to each taxon. These values are generated by subtracting 1 from the p-values generated by the ISA, with values closer to 1 indicating taxa that are more specific. The analysis is based on the ISA described by Dufrêne & Legendre (1997) and the permutation test is based on that described by McCune & Grace (2002). The taxon_specificity_index output provides a list of these taxa level indices.

Community level specificity indices are then generated through CWM analysis using taxon level specificity indices. Values closer to one indicate communities that are more host specific. This output can be found in the community_specificity_index output.

The user has the option of removing rare taxa from the CWM analysis. Rare taxa bias communities to appear more host specific. The rare taxa removal threshold is determined by first grouping taxa by the number of samples in which they occur, then removing the groups in which the taxa are so rare that no taxon has a significant result from the ISA. The default is for rare taxa to be removed.
}
\examples{
out <- specifiR_physeq(physeq = ps,groups = "depth",n.perm = 99)
out$community_specificity_index

}
