% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifiR.R
\name{specifiR}
\alias{specifiR}
\title{Quantify group specificity of entire communities for each sample}
\usage{
specifiR(
  comm,
  groups,
  seed = 666,
  n.perm = 999,
  pval.cutoff = 0.05,
  max.ratio = 0,
  ovp.plot = FALSE
)
}
\arguments{
\item{comm}{Data frame. Data frame (or coercible to "data.frame") of community abundance values. Rows are samples, columns are taxa. Values are raw observation numbers.}

\item{groups}{Character. Vector of grouping levels. Must be same length as number of samples (rows) of comm.}

\item{seed}{Numeric of length 1. Random seed for reproducibility. Default = 666.}

\item{n.perm}{Positive numeric (integer) of length 1. Number of permutations for Monte Carlo permutation test. Default = 999.}

\item{pval.cutoff}{Positive numeric of length 1, between 0 and 1. The P-value cutoff for significance. Default = 0.05.}

\item{max.ratio}{The maximum ratio of significant:insignificant P-values within a group to indicate removal. Taxa with fidelity to groups at or less than this value will be removed. Only the first N occurence groups that have this value or lower will be flagged for taxon removal. Default = 0. You are unlikely to want to change this value.}

\item{ovp.plot}{Logical. Should a plot of occupancy vs. p-values be generated? Default = FALSE.}
}
\value{
Named List. This returns a list with 3 data frame elements: community_specificity_index = The main result showing community weighted mean indicator values for each sample; taxon_specificity_index = Intermediate result (`comm_name` `iv.max` `p.value` `occurrence` `occurrence_groups` `taxon_index`); isa_results = Intermediate results. taxon-level indicator species analysis with each given group level, and the indicator results for each taxon.
}
\description{
Abbey describe what the method does
Subsequence lines for long description
}
\details{
This function implements the "Community Weighted Mean Indicator" Analysis for quantifying group specificity of entire communities.
}
\examples{
# make mock community and grouping data
comm_matrix <- matrix(rpois(30 * 100, lambda = 5),nrow = 30,ncol = 100)
colnames(comm_matrix) <- paste0("taxon.", 1:100)
comm_df <- as.data.frame(comm_matrix)
groups <- factor(rep(paste0("Group", 1:3), length.out = 100))

# run specifiR with default settings
out <- specifiR(comm = comm_df, groups = groups)
out$community_specificity_index

}
