% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifiR_physeq.R
\name{specifiR_physeq}
\alias{specifiR_physeq}
\title{Quantify group specificity of entire communities for each sample}
\usage{
specifiR_physeq(
  physeq,
  groups,
  seed = 666,
  n.perm = 999,
  pval.cutoff = 0.05,
  max.ratio = 0,
  ovp.plot = FALSE
)
}
\arguments{
\item{physeq}{Phyloseq object. Must contain a valid "otu_table" slot containing raw observation counts, and a valid "sample_data" slot containing the grouping variable.}

\item{groups}{Character vector of length 1. The quoted name of the sample_data column that has grouping information.}

\item{seed}{Numeric of length 1. Random seed for reproducibility. Default = 666.}

\item{n.perm}{Positive numeric (integer) of length 1. Number of permutations for Monte Carlo permutation test. Default = 999.}

\item{pval.cutoff}{Positive numeric of length 1, between 0 and 1. The P-value cutoff for significance. Default = 0.05.}

\item{max.ratio}{The maximum ratio of significant:insignificant P-values within a group to indicate removal. Taxa with fidelity to groups at or less than this value will be removed. Only the first N occurence groups that have this value or lower will be flagged for taxon removal. Default = 0. You are unlikely to want to change this value.}

\item{ovp.plot}{Logical. Should a plot of occupancy vs. p-values be generated? Default = FALSE.}
}
\value{
Named List. This returns a list with 3 data frame elements: community_specificity_index = The main result showing community weighted mean indicator values for each sample; taxon_specificity_index = Intermediate result (`comm_name` `iv.max` `p.value` `occurrence` `occurrence_groups` `taxon_index`); isa_results = Intermediate results. taxon-level indicator species analysis with each given group level, and the indicator results for each taxon.
}
\description{
Abbey describe what the method does
Subsequent lines for long description.
This version of the package is a convenience wrapper for phyloseq users.
}
\details{
This function implements the "Community Weighted Mean Indicator" Analysis for quantifying group specificity of entire communities.
}
\examples{
out <- specifiR_physeq(physeq = ps,groups = "depth",n.perm = 99)
out$community_specificity_index

}
